+++ internal/tui/pages/tunnels/tunnels.go
@@ -2,7 +2,6 @@ package tunnels
 
 import (
 	"github.com/charmbracelet/bubbles/list"
-	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 )